Fix UB in `Arc::<[MaybeUninit<T>]>::as_mut_slice`
